requirements.txt
flask
python-docx
anthropic
python-dotenv
lxml
requests
